삭제하기 동의

+++ src/pages/Home.jsx
@@ -6,7 +6,8 @@ export default function Home() {
   const [users, setUsers] = useState([]); //유저객체 배열
   // 삭제하기 함수
   const deleteUser = async (id) => {
-    await axios.delete(`http://localhost:8080/users/${id}`);
+    if (window.confirm("정말로 삭제 하겠ㅅ브니까?"))
+      await axios.delete(`http://localhost:8080/users/${id}`);
     loadUsers(); //삭제후에 다시 삭제된 개 유저배열을 받음
   };
   //백엔드에서 유저들을 가져오는 함수
